psycheos-interpreter/requirements.txt
# Core dependencies
anthropic>=0.18.0
aiogram>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional but useful
structlog>=23.0.0
